fix: type mismatch & geolocator 13 API compat
- app_constants: ridePrice/rideCommission changed from int to double
  to match setCalculatedPrice(double) and RideSummaryCard.price (double)
- location_service: removed locationSettings named parameter from
  getCurrentPosition and getPositionStream (not supported in
  geolocator ^13.0.0); use desiredAccuracy/timeLimit/distanceFilter
  directly instead

+++ lib/core/constants/app_constants.dart
@@ -8,9 +8,9 @@ class AppConstants {
   static const String currency = 'FDJ';
 
   // Ride Pricing
-  static const int ridePrice = 250;
-  static const int rideCommission = 125;
-  static const int driverEarning = ridePrice - rideCommission;
+  static const double ridePrice = 250.0;
+  static const double rideCommission = 125.0;
+  static const double driverEarning = ridePrice - rideCommission;
 
   // Phone
   static const String countryCode = '+253';
